feat: #526 organize FilePath
