+++ packages/liferay-js-toolkit-core/src/project/Jar.ts
@@ -7,7 +7,7 @@ import prop from 'dot-prop';
 import fs from 'fs-extra';
 import readJsonSync from 'read-json-sync';
 
-import FilePath from '../file-path';
+import FilePath from '../file/FilePath';
 import {print, warn} from '../format';
 import Project from './Project';
 import {getFeaturesFilePath} from './getFeaturesFilePath';
